# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gradway Education Consultancy API",
    "description": "Machine-readable API for Gradway (Pvt) Ltd - Colombo, Sri Lanka study abroad consultancy services, program evaluations, and inquiry verification.",
    "version": "1.0.0",
    "contact": {
      "name": "Gradway Admissions Team",
      "email": "info@gradwayedu.com",
      "url": "https://gradwayedu.com"
    }
  },
  "servers": [
    {
      "url": "https://gradwayedu.com",
      "description": "Production Server"
    }
  ],
  "paths": {
...
    "/api/upload/validate": {
      "post": {
        "summary": "Validate Academic Document Upload",
        "description": "Validates binary file signatures, size boundaries, and MIME types for academic transcripts, SOPs, and financial docs.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fileName", "mimeType", "base64Content"],
                "properties": {
                  "fileName": { "type": "string", "example": "academic_transcript.pdf" },
                  "mimeType": { "type": "string", "example": "application/pdf" },
                  "base64Content": { "type": "string", "description": "Base64-encoded document data" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Document validated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "valid": { "type": "boolean", "example": true },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed due to invalid signature or size"
          }
        }
      }
    },
...
  }
}